+++ src/DIPlib_sources.cmake
@@ -64,6 +64,7 @@ analysis/distance_distribution.cpp
 analysis/findshift.cpp
 analysis/fractal_dimension.cpp
 analysis/granulometry.cpp
+analysis/monogenic_signal.cpp
 analysis/pixel_pair_sampling.cpp
 analysis/structure_tensor.cpp
 analysis/subpixel_location.cpp